catalog  explore section done

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require("../models/SubSection");
 // CREATE a new section
 exports.createSection = async (req, res) => {
 	try {
@@ -54,7 +55,7 @@ exports.createSection = async (req, res) => {
 // UPDATE a section
 exports.updateSection = async (req, res) => {
 	try {
-		const { sectionName, sectionId } = req.body;
+		const { sectionName, sectionId, courseId } = req.body;
 		// validation
 		if(!sectionName || !sectionId)
 		{
@@ -69,10 +70,19 @@ exports.updateSection = async (req, res) => {
 			{ sectionName }, // change the name
 			{ new: true }
 		);
+
+		const course = await Course.findById(courseId)
+      .populate({
+        path: "courseContent",
+        populate: {
+          path: "subSection",
+        },
+      })
+      .exec()
 		res.status(200).json({
 			success: true,
-			message: "Section Updated Successfully",
-			section,
+			message: section,
+			data: course,
 		});
 	} catch (error) {
 		console.error("Error updating section:", error);
@@ -86,20 +96,45 @@ exports.updateSection = async (req, res) => {
 // DELETE a section 
 exports.deleteSection = async (req, res) => {
 	try {
-		// assuming id is sent in params
-		const { sectionId } = req.params;
-		await Section.findByIdAndDelete(sectionId);
-		res.status(200).json({
-			success: true,
-			message: "Section deleted",
-		});
-
-		// remaining: delete from secions id from course schema
-	} catch (error) {
-		console.error("Error deleting section:", error);
-		res.status(500).json({
+		const { sectionId, courseId } = req.body
+		await Course.findByIdAndUpdate(courseId, {
+		  $pull: {
+			courseContent: sectionId,
+		  },
+		})
+		const section = await Section.findById(sectionId)
+		if (!section) {
+		  return res.status(404).json({
 			success: false,
-			message: "Internal server error",
-		});
-	}
+			message: "Section not found",
+		  })
+		}
+		// Delete the associated subsections
+		await SubSection.deleteMany({ _id: { $in: section.subSection } })
+	
+		await Section.findByIdAndDelete(sectionId)
+	
+		// find the updated course and return it
+		const course = await Course.findById(courseId)
+		  .populate({
+			path: "courseContent",
+			populate: {
+			  path: "subSection",
+			},
+		  })
+		  .exec()
+	
+		res.status(200).json({
+		  success: true,
+		  message: "Section deleted",
+		  data: course,
+		})
+	  } catch (error) {
+		console.error("Error deleting section:", error)
+		res.status(500).json({
+		  success: false,
+		  message: "Internal server error",
+		  error: error.message,
+		})
+	  }
 };
